+++ internal/metrics/cluster_metrics.go
@@ -1,7 +1,7 @@
-// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Greenhouse contributors
+// SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Greenhouse contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package cluster
+package metrics
 
 import (
 	"time"
@@ -14,20 +14,20 @@ import (
 )
 
 var (
-	clusterReadyGauge = prometheus.NewGaugeVec(
+	ClusterReadyGauge = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Name: "greenhouse_cluster_ready",
 			Help: "Indicates whether the cluster is ready",
 		},
 		[]string{"cluster", "namespace", "owned_by"})
 
-	kubernetesVersionsGauge = prometheus.NewGaugeVec(
+	KubernetesVersionsGauge = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Name: "greenhouse_cluster_k8s_versions_total",
 		},
 		[]string{"cluster", "namespace", "version", "owned_by"})
 
-	secondsToTokenExpiryGauge = prometheus.NewGaugeVec(
+	SecondsToTokenExpiryGauge = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Name: "greenhouse_cluster_kubeconfig_validity_seconds",
 		},
@@ -35,18 +35,19 @@ var (
 )
 
 func init() {
-	metrics.Registry.MustRegister(kubernetesVersionsGauge)
-	metrics.Registry.MustRegister(secondsToTokenExpiryGauge)
+	metrics.Registry.MustRegister(KubernetesVersionsGauge)
+	metrics.Registry.MustRegister(SecondsToTokenExpiryGauge)
+	metrics.Registry.MustRegister(ownedByLabelMissingGauge)
 }
 
-func updateMetrics(cluster *greenhousev1alpha1.Cluster) {
+func UpdateClusterMetrics(cluster *greenhousev1alpha1.Cluster) {
 	kubernetesVersionLabels := prometheus.Labels{
 		"cluster":   cluster.Name,
 		"namespace": cluster.Namespace,
 		"version":   cluster.Status.KubernetesVersion,
 		"owned_by":  cluster.Labels[greenhouseapis.LabelKeyOwnedBy],
 	}
-	kubernetesVersionsGauge.With(kubernetesVersionLabels).Set(float64(1))
+	KubernetesVersionsGauge.With(kubernetesVersionLabels).Set(float64(1))
 
 	secondsToExpiry := cluster.Status.BearerTokenExpirationTimestamp.Unix() - time.Now().Unix()
 	secondsToExpiryLabels := prometheus.Labels{
@@ -54,7 +55,7 @@ func updateMetrics(cluster *greenhousev1alpha1.Cluster) {
 		"namespace": cluster.Namespace,
 		"owned_by":  cluster.Labels[greenhouseapis.LabelKeyOwnedBy],
 	}
-	secondsToTokenExpiryGauge.With(secondsToExpiryLabels).Set(float64(secondsToExpiry))
+	SecondsToTokenExpiryGauge.With(secondsToExpiryLabels).Set(float64(secondsToExpiry))
 
 	clusterReadyLabels := prometheus.Labels{
 		"cluster":   cluster.Name,
@@ -62,8 +63,8 @@ func updateMetrics(cluster *greenhousev1alpha1.Cluster) {
 		"owned_by":  cluster.Labels[greenhouseapis.LabelKeyOwnedBy],
 	}
 	if cluster.Status.IsReadyTrue() {
-		clusterReadyGauge.With(clusterReadyLabels).Set(float64(1))
+		ClusterReadyGauge.With(clusterReadyLabels).Set(float64(1))
 	} else {
-		clusterReadyGauge.With(clusterReadyLabels).Set(float64(0))
+		ClusterReadyGauge.With(clusterReadyLabels).Set(float64(0))
 	}
 }
